% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{calc.se}
\alias{calc.se}
\title{calc.se}
\usage{
calc.se(x, t.max)
}
\arguments{
\item{x}{the vector of CTD data e.g. oxygen) by depth bins}

\item{t.max}{the maximum autocorrelation interval to use, should be sqrt(length(x))}
}
\value{
standard error of the mean
}
\description{
calculate the standard error from CTD casts
}
\examples{

}
