% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_loo.R
\name{compare_loo}
\alias{compare_loo}
\title{compare_loo}
\usage{
compare_loo(species, data.2.use)
}
\arguments{
\item{species}{what species to use.  either "zoop", "fish", "Pacific Hake", or "Pacific Herring"}

\item{data.2.use}{what response variable to use either "Sa", "fullness", or "dietcomp"}
}
\value{
list(comp, stacking_wts)
}
\description{
use leave-one-out cross validation approximation for a collection of n models
and calculate the difference in expected log point density (and se), and also
calculate model weights.  Stan output must already exist
}
\examples{



}
