% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctd_data}
\alias{ctd_data}
\title{ctd_data}
\format{
a large list that takes a matlab structure object giving ctd data in 1m depth bins
\describe{
 \item{hc}{the element of the list that matters: contains depth (m) first element, temperature in 2nd element , dissolved oxygen in the 16th element, and 14th has all of the sample information}
}
}
\usage{
ctd_data
}
\description{
ctd_data
}
\keyword{datasets}
