% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euphausiid_lengths}
\alias{euphausiid_lengths}
\title{euphausiid_lengths}
\format{
a list, with names elements corresponding to individual multinet tows.  Each sample is an measured euphausiid length
}
\usage{
euphausiid_lengths
}
\description{
euphausiid_lengths
}
\keyword{datasets}
