% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feeding_data}
\alias{feeding_data}
\title{feeding_data}
\format{
a two element list that has all of the measured feeding metrics for each species
\describe{
 \item{ci}{the fraction of diet that is euphauiids}
 \item{ci_SE}{estimated standard error of diet fraction}
 \item{logfullness}{ratio of stomach contents to fish length^3}
 \item{selogfullness}{standard error of ratio of stomach contents to fish length^3}
 \item{N}{samplesize}
 \item{logit.ci}{logit of diet fraction}
 \item{logit.ci.se}{standard error of logit of diet fraction}
}
}
\usage{
feeding_data
}
\description{
feeding_data
}
\keyword{datasets}
