% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_backscatter}
\alias{fish_backscatter}
\title{fish_backscatter}
\format{
a three element list that has all of the measured feeding metrics for combined fish, Pacific Herring and Pacific Hake
\describe{
 \item{Year}{Year}
 \item{Month}{Month}
 \item{Diel}{day or night sample?}
 \item{A}{site A, Union}
 \item{B}{site B, Hoodsport}
 \item{C}{site C, Duckabush}
 \item{D}{site D, Dabob Bay}
}
}
\usage{
fish_backscatter
}
\description{
fish_backscatter
}
\keyword{datasets}
