% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{fixed.effect}
\alias{fixed.effect}
\title{fixed.effect}
\usage{
fixed.effect(x, se)
}
\arguments{
\item{x}{the vector of means}

\item{se}{the vector of standard errors}
}
\value{
fixed effect mean and standard error: c(mean, standard error)
}
\description{
calculate the average (and SE) of multiple means (x1, x2,...xn) each with corresponding standard error
}
\examples{

}
