% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDOdataFN.R
\name{formatDOdataFN}
\alias{formatDOdataFN}
\title{formatDOdataFN}
\usage{
formatDOdataFN()
}
\value{
a dataframe containing year, diel period, site, month, and mean DO (<20 m) for each
}
\description{
retrieve dissolved oxygen data from matlab file HCmat, and format it for use in modeling
}
\examples{

}
