% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_posterior_predation.R
\name{generate_posterior_predation}
\alias{generate_posterior_predation}
\title{generate_posterior_prediction}
\usage{
generate_posterior_predation()
}
\value{
a plot in the active plot window, with median listed and 80% credibility interval shown
}
\description{
uses the posterior of zooplankton and herring backscatter
and herring feeding, to show posterior probability of predation intensity
}
\examples{

}
