% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{get.ci}
\alias{get.ci}
\title{calc.ci}
\usage{
get.ci(x, y, value)
}
\arguments{
\item{x}{the parameter values}

\item{y}{the probability densities}

\item{value}{what percentile range should the credibility interval span (usually 0.8)}
}
\value{
c(lower, upper, lower density, upper density)
}
\description{
calculate the X% credibility interval from kernel-smoothed MCMC output
}
\examples{

}
