% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{get.percentile}
\alias{get.percentile}
\title{get.percentile}
\usage{
get.percentile(x, y, value)
}
\arguments{
\item{x}{the vector parameter values}

\item{y}{the vector probability densities}

\item{value}{the percentile to return (between 0 and 1)}
}
\value{
the estimated percentile from the posterior probability
}
\description{
finds a given percentile from kernel-smoothed MCMC chains
}
\examples{

}
