% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_euphausiid_lengths.R
\name{get_euphausiid_lengths}
\alias{get_euphausiid_lengths}
\title{get_euphausiid_lengths}
\usage{
get_euphausiid_lengths()
}
\value{
a data frame listing year, month, site, diel and the proportion euphausiids that are large
}
\description{
calculates the fraction of euphausiids that are greater than 10 mm, for each sampling event
}
\examples{

}
