% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_euphausiid_numbers.R
\name{get_euphausiid_numbers}
\alias{get_euphausiid_numbers}
\title{get_euphausiid_numbers}
\usage{
get_euphausiid_numbers()
}
\value{
a data matrix given the euphausiid numerical proportion for each sampling event, and the proportion of NASC attributable to euphausiids
}
\description{
use the multi-net tow data to estimate the fraction of NASC likely attributable to euphausiids.
}
\examples{



}
