% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Zoop_numbers_proportions.R
\name{get_Zoop_numbers_proportions}
\alias{get_Zoop_numbers_proportions}
\title{get_Zoop_numbers_proportions}
\usage{
get_Zoop_numbers_proportions(species)
}
\arguments{
\item{species}{which zooplankton species or species group should be used.  Must match one of the column
names in data:zoopcomp}
}
\value{
a vector containing estimated numerical proportions for each taxa
}
\description{
calculate numerical proportions of zooplankton sampled in multi-nets for multiple taxa
}
\examples{

}
