% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{inv.logit}
\alias{inv.logit}
\title{inv.logit}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{a real number, or vector of numbers}
}
\value{
exp(x) / (1 + exp(x))
}
\description{
calculate the inverse logit of a real number
}
\examples{

}
