% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dietdata.R
\name{load_dietdata}
\alias{load_dietdata}
\title{load_dietdata}
\usage{
load_dietdata(species, data.2.use)
}
\arguments{
\item{species}{what species to use.  either "Pacific Hake", or "Pacific Herring"}

\item{data.2.use}{what metric to extract, either "dietcomp" or "fullness"}
}
\value{

}
\description{
extracts either stomach fullness and diet composition (fraction of diets that were euphausiids)
for each sampling event, returns a data frame
}
\examples{

}
