% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sa_fish_data.R
\name{load_sa_fish_data}
\alias{load_sa_fish_data}
\title{load_sa_fish_data}
\usage{
load_sa_fish_data(species)
}
\arguments{
\item{species}{what species to use.  either  "fish", "Pacific Hake", or "Pacific Herring"}
}
\value{
A dataframe containing year, month, site, diel period, and column y for the response
variable (backscatter), and mean dissolved oxygen.  Column se is set to 0
}
\description{
extracts either the acoustic backscatter from specified species / groups
}
\examples{

}
