% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_zoop_data.R
\name{load_zoop_data}
\alias{load_zoop_data}
\title{load_sa_fish_data}
\usage{
load_zoop_data()
}
\value{
A dataframe containing year, month, site, diel period, and column y for the response
variable (backscatter), and mean dissolved oxygen.  Column se is set to 0
}
\description{
extracts the acoustic backscatter for zooplankton
}
\examples{

}
