% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{logit.mod.se}
\alias{logit.mod.se}
\title{logit.mod.se}
\usage{
logit.mod.se(x)
}
\arguments{
\item{x}{the standard error of non transformed data}
}
\value{
1 / (x * (1 - x))
}
\description{
calculate standard error of a logit transform
}
\examples{

}
