% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFUNs.R
\name{logit}
\alias{logit}
\title{logit}
\usage{
logit(x)
}
\arguments{
\item{x}{a real number, or vector of numbers between 0 and 1}
}
\value{
log(p/(1-p))
}
\description{
calculate the logit of a probability, p
}
\examples{

}
