% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netdepths}
\alias{netdepths}
\title{netdepths}
\format{
a dataframe containing lower and upper bound for each multinet tow
\describe{
 \item{Net}{alphanumeric code for tow file}
 \item{lower}{lower depth}
 \item{upper}{upper depth}
}
}
\usage{
netdepths
}
\description{
netdepths
}
\keyword{datasets}
