% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_backscatter_posteriors.R
\name{plot_backscatter_posteriors}
\alias{plot_backscatter_posteriors}
\title{plot_backscatter_posteriors}
\usage{
plot_backscatter_posteriors()
}
\value{
stores a pdf file "acoustic_posterior" in graphics folder
}
\description{
creates a plot showing posterior probability distributions for backscatter metrics
}
\examples{

}
