% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dietcomp_data.R
\name{plot_dietcomp_data}
\alias{plot_dietcomp_data}
\title{plot_dietcomp_data}
\usage{
plot_dietcomp_data()
}
\value{
stores a pdf file "dietcomp.pdf" in graphics folder
}
\description{
creates a plot showing the diet fraction of herring and hake, by month, site, and time period
for each year
}
\examples{

}
