% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_do_profiles.R
\name{plot_do_profiles}
\alias{plot_do_profiles}
\title{plot_do_profiles}
\usage{
plot_do_profiles()
}
\value{
stores a pdf file "doprofiles.pdf" in graphics folder
}
\description{
creates a plot showing dissolved oxygen profiles (figure S3)
}
\examples{

}
