% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feeding_posteriors.R
\name{plot_feeding_posteriors}
\alias{plot_feeding_posteriors}
\title{plot_backscatter_posteriors}
\usage{
plot_feeding_posteriors()
}
\value{
stores a pdf file "feeding_posterior" in graphics folder
}
\description{
creates a plot showing posterior probability distributions for feeding metrics
}
\examples{

}
