% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fish_backscatter_data.R
\name{plot_fish_backscatter_data}
\alias{plot_fish_backscatter_data}
\title{plot_fish_backscatter_data}
\usage{
plot_fish_backscatter_data()
}
\value{
stores a pdf file "fish_sa" in graphics folder
}
\description{
creates a plot showing the estimated backscatter for Pacific Hake and Pacific Herring
By month, sample site, and year.
}
\examples{

}
