% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fraction_large_euphausiid.R
\name{plot_fraction_large_euphausiids}
\alias{plot_fraction_large_euphausiids}
\title{plot_fraction_large_euphausiids}
\usage{
plot_fraction_large_euphausiids()
}
\value{
stores a pdf file "fraction_large_euphausiid" in graphics folder
}
\description{
creates a plot showing the estimated proportion of euphausiids that were larger than 10 mm,
accross seasons, sites and years. Appears as Figure S5
}
\examples{

}
