% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fullness_data.R
\name{plot_fullness_data}
\alias{plot_fullness_data}
\title{plot_fish_backscatter_data}
\usage{
plot_fullness_data()
}
\value{
stores a pdf file "fish_sa" in graphics folder
}
\description{
creates a plot showing the stomach fullness for Pacific Hake and Pacific Herring
By month, sample site, and year.  This is figure 4.
}
\examples{

}
