% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot meanDO by Site.R
\name{plot_meanDO_by_site}
\alias{plot_meanDO_by_site}
\title{plot_meanDO_by_site}
\usage{
plot_meanDO_by_site()
}
\value{
stores a pdf file "meanDObysite" in graphics folder
}
\description{
creates 2 x 2 panel plot of mean DO  by month for each site
}
\examples{

}
