% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot pnasc vs nasc.R
\name{plot_pnasc_vs_nasc}
\alias{plot_pnasc_vs_nasc}
\title{plot_pnasc_vs_nasc}
\usage{
plot_pnasc_vs_nasc()
}
\value{
stores a pdf file "pnascvsnasc.pdf" in graphics folder
}
\description{
creates a plot showing relationship between pnasc and log(nasc)
}
\examples{

}
