% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_profiles.R
\name{plot_temp_profiles}
\alias{plot_temp_profiles}
\title{plot_temp_profiles}
\usage{
plot_temp_profiles()
}
\value{
stores a pdf file "tprofiles.pdf" in graphics folder
}
\description{
creates a plot showing temperature profiles (figure S4)
}
\examples{

}
