% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoop_backscatter_data.R
\name{plot_zoop_backscatter_data}
\alias{plot_zoop_backscatter_data}
\title{plot_zoop_backscatter_data}
\usage{
plot_zoop_backscatter_data()
}
\value{
stores a pdf file "zoop_by_month" in graphics folder
}
\description{
creates a plot showing the estimated backscatter for zooplankton
By month, sample site, and year.
}
\examples{

}
