% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run All Bayesian Models.R
\name{run_all_bayesian_models}
\alias{run_all_bayesian_models}
\title{run_all_bayesian_models}
\usage{
run_all_bayesian_models(species, data.2.use)
}
\arguments{
\item{species}{what species to use.  either  "zoop", "fish", "Pacific Hake", or "Pacific Herring"}

\item{data.2.use}{what response metric to use.  Either "Sa", "dietcomp", or "fullness". The latter
two can only be used if species is Pacific Hake or Pacific Herring}
}
\value{
stores .Rdata objects containing lists of outputs for each model, stored in outputs/modelselection/ folders
}
\description{
fits all candidate bayesian models, saves the MCMC output and a namedd list with class c("psisi_loo")
}
\examples{

}
