% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoop_backscatter}
\alias{zoop_backscatter}
\title{zoop_backscatter}
\format{
a data frame containing acoustic backscatter classified as zooplankton, for each sampling event
}
\usage{
zoop_backscatter
}
\description{
zoop_backscatter
}
\keyword{datasets}
