% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoopcomp}
\alias{zoopcomp}
\title{zoopcomp}
\format{
a dataframe containing counts of zooplankton taxa for each multinet tow
\describe{
 \item{Year}{year of sample}
 \item{Month}{numeric version of month}
 \item{Day}{day of sample}
 \item{Site}{site, listed in 2 letter abbreviations, UN = Union, HP = hoodsport, DU = duckabush, DB = Dabob bay}
 \item{Diel}{day or night sample}
 \item{Filename}{original filename, designates alpha-numeric multinet tow ID}
 \item{species names}{several columns indicating individual taxa}
 \item{Monthname}{three character month abbreviations, matches what is used in other datasets}
 \item{Sitename}{turn Site into A, B, C, D to match other datasets}
}
}
\usage{
zoopcomp
}
\description{
zoopcomp
}
\keyword{datasets}
